Tracking: Cleanup (#269)
* Tracking: Simplify components

* Tracking: API cleanup

* FIX: Compile issues

* DELETE: Unnecessary components

* FORMAT: Lint fixes

* DELETE: Unnecessary code

* TRACKING: Cleanup / simplify components

* Tracking: Fix Recording type params

* MOVE: Reads/Writes -> Tracker

* FIX: Using Tracker for cache

* Tests: Fixed unit-test compilation issues

* FIX: Compilation issues in tests

* FIX: Unit-tests

* Build: Setting tracking to stable version number

+++ src/lib/kotlin/slatekit-tracking/src/main/kotlin/slatekit/tracking/Expiry.kt
@@ -4,13 +4,10 @@ import slatekit.common.DateTime
 
 data class Expiry(val seconds:Long,
                   val started:DateTime = DateTime.now(),
-                  val expires:DateTime = started.plusSeconds(seconds.toLong())) {
+                  val expires:DateTime = started.plusSeconds(seconds)) {
 
     fun isExpired(): Boolean {
-        return if (seconds <= 0)
-            false
-        else
-            expires <= DateTime.now()
+        return if (seconds <= 0) false else expires <= DateTime.now()
     }
